backend/requirements.txt
# FastAPI and server
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Machine Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python>=4.7.0
Pillow>=9.5.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0